pjproject: update to last stable release (1.10)
Remove some stuff we don't need:
    third_party libs
    Visual Studio / Xcode files
    symbian files

    pjmedia-audiodev (and update some source files accordingly)
    pjmedia-codec

    tests, documentation

define PJMEDIA_RESAMPLE_IMP PJMEDIA_RESAMPLE_NONE
    => so we don't rely on resample library

define PJMEDIA_HAS_SRTP     0
    => so we don't rely on srtp library

We do resampling and srtp ourselves anyway

+++ daemon/libs/pjproject/pjmedia/src/pjmedia/wsola.c
@@ -1,6 +1,6 @@
-/* $Id: wsola.c 2850 2009-08-01 09:20:59Z bennylp $ */
+/* $Id: wsola.c 3553 2011-05-05 06:14:19Z nanang $ */
 /* 
- * Copyright (C) 2008-2009 Teluu Inc. (http://www.teluu.com)
+ * Copyright (C) 2008-2011 Teluu Inc. (http://www.teluu.com)
  * Copyright (C) 2003-2008 Benny Prijono <benny@prijono.org>
  *
  * This program is free software; you can redistribute it and/or modify
@@ -16,17 +16,6 @@
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA 
- *
- *  Additional permission under GNU GPL version 3 section 7:
- *
- *  If you modify this program, or any covered work, by linking or
- *  combining it with the OpenSSL project's OpenSSL library (or a
- *  modified version of that library), containing parts covered by the
- *  terms of the OpenSSL or SSLeay licenses, Teluu Inc. (http://www.teluu.com)
- *  grants you additional permission to convey the resulting work.
- *  Corresponding Source for a non-source form of such a combination
- *  shall include the source code for the parts of OpenSSL used as well
- *  as that of the covered work.
  */
 #include <pjmedia/wsola.h>
 #include <pjmedia/circbuf.h>
@@ -861,7 +850,7 @@ PJ_DEF(pj_status_t) pjmedia_wsola_save( pjmedia_wsola *wsola,
 	    /* Continue applying fade out to the extra samples */
 	    if (reg2_len == 0) {
 		wsola_fade_out(wsola, reg1 + reg1_len - count, count);
-	    } else if ((int)reg2_len >= count) {
+	    } else if (reg2_len >= count) {
 		wsola_fade_out(wsola, reg2 + reg2_len - count, count);
 	    } else {
 		unsigned tmp = count - reg2_len;
